% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{autoplot_fun}
\alias{autoplot_fun}
\title{autoplot_fun}
\usage{
autoplot_fun(x, y, plot.tag.position = c(0, 0.95))
}
\arguments{
\item{x}{A fitdists object as returned by ssdtools.}

\item{y}{labels to plot on each grob.}

\item{plot.tag.position}{Tag position for labels.}
}
\value{
A ggplot2 grob of the fitdists object.
}
\description{
autoplot_fun
}
